+++ aws-sdk-core/lib/aws-sdk-core/applicationdiscoveryservice.rb
@@ -1,0 +1,5 @@
+Aws.add_service(:ApplicationDiscoveryService, {
+  api: "#{Aws::API_DIR}/discovery/2015-11-01/api-2.json",
+  docs: "#{Aws::API_DIR}/discovery/2015-11-01/docs-2.json",
+  examples: "#{Aws::API_DIR}/discovery/2015-11-01/examples-1.json",
+})